+++ src/test/converter2/issues/gh2106.ts
@@ -1,0 +1,17 @@
+export function balance(address: string): Coin {
+    return {
+        amount: "",
+        denom: "",
+    };
+}
+
+export interface Coin {
+    denom: string;
+    amount: string;
+}
+
+export const Coin = {
+    thisIsAValue: true,
+};
+
+export type TypeOf = typeof Coin;