+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Bug Fixes
 
+-   If both an interface and a variable share a name/symbol, TypeDoc will no longer link to the variable when referenced in a type position, #2106.
 -   `notDocumented` validation will no longer require documentation for data within parameters that cannot be documented via `@param`, #2291.
 -   "defined in" locations for signatures will now always be contained within the function declaration's location. This prevents defined in sometimes pointing to node_modules, #2307.
 -   Type parameters will now be resolved for arrow-methods on classes like regular class methods, #2320.